Cleaned up tests, verified to pass on all supported Python versions with tox
